+++ Examples/Examples/ESTViewController.m
@@ -8,12 +8,14 @@
 
 #import "ESTViewController.h"
 #import "ESTBeaconTableVC.h"
+#import "ESTCloudBeaconTableVC.h"
 #import "ESTDistanceDemoVC.h"
 #import "ESTProximityDemoVC.h"
 #import "ESTNotificationDemoVC.h"
 #import "ESTTemperatureDemoVC.h"
 #import "ESTAccelerometerDemoVC.h"
 #import "ESTUpdateFirmwareDemoVC.h"
+#import "ESTCloudAuthVC.h"
 
 @interface ESTDemoTableViewCell : UITableViewCell
 
@@ -47,10 +49,23 @@
     self.tableView.sectionHeaderHeight = 20;
     [self.tableView registerClass:[ESTDemoTableViewCell class] forCellReuseIdentifier:@"DemoCellIdentifier"];
     
+    UIBarButtonItem *authButton = [[UIBarButtonItem alloc] initWithTitle:@"Authorize"
+                                                                   style:UIBarButtonItemStylePlain
+                                                                  target:self
+                                                                  action:@selector(authorizeBtnTapped:)];
+    self.navigationItem.rightBarButtonItem = authButton;
+    
     self.beaconDemoList = @[ @[@"Distance Demo", @"Proximity Demo",@"Notification Demo"],
                              @[@"Temperature Demo", @"Accelerometer Demo"],
-                             @[@"Update Firmware Demo"]];
+                             @[@"Update Firmware Demo", @"My beacons in Cloud Demo"]];
 }
+
+-(void)authorizeBtnTapped:(UIButton *)button
+{
+    ESTCloudAuthVC *cloudVC = [ESTCloudAuthVC new];
+    [self.navigationController pushViewController:cloudVC animated:YES];
+}
+
 
 #pragma mark - Table view data source
 
@@ -178,6 +193,12 @@
                     ESTUpdateFirmwareDemoVC *updateFirmwareVC = [[ESTUpdateFirmwareDemoVC alloc] initWithBeacon:beacon];
                     [self.navigationController pushViewController:updateFirmwareVC animated:YES];
                 }];
+                break;
+            }
+            case 1:
+            {
+                demoViewController = [ESTCloudBeaconTableVC new];
+                
                 break;
             }
             default: